Switching Project Over to WPF
Switching the project over from a console application to a wpf application.
This is done to add wpf windows as GUIs. The Main method has been custom defined in the program.cs class, this way the user is able to use the console application chatbot as well as open GUIs for the new features like the task manager and reminder feature. This way ensures the application retains its original design and content but allows the user to use GUI features in the application as well. The GUIs can be launched from the console application by the user by way of a menu.